+++ src/node/utils.py
@@ -29,6 +29,12 @@ class Unset(object):
     def __repr__(self):
         return '<UNSET>'
 
+    def __copy__(self):
+        return self
+
+    def __deepcopy__(self, memo):
+        return self
+
 
 UNSET = Unset()